Started implementation of deck building
Added a "Selected Deck" Character Deck to game mode, in hopes that it will be a bit more intuitive moving forward. Currently, an error is thrown that nothing is being read when trying to pull information... despite information being pulled. Probably a simply mistake, game technically works, but I'll try to figure out why it is throwing that error