Implemented ability to add cards to a deck
Very basic ability to had a card to a deck. Main issue is how we want to handle cards in the future. Should each card be it's own Actor? One general actor class that we manually enter in the info for when ever the card is added to a deck?